+++ subterra-engine/src/main/java/io/toterra/subterra/engine/worldgen/profiler/facade/ApiRegistry.java
@@ -1,4 +1,4 @@
-package io.toterra.subterra.profiler.facade;
+package io.toterra.subterra.engine.worldgen.profiler.facade;
 
 import java.io.IOException;
 import java.nio.file.Path;